refactor(sorting): move out of text plugin
