+++ apps/web/src/components/taxonomy/taxonomy-move-copy/get-preview-string-from-exercise.tsx
@@ -1,11 +1,14 @@
-import { extractStringFromRowsTextAndMultimedia } from '@editor/plugins/text/utils/static-extract-text'
+import { extractStringFromTextDocument } from '@editor/plugins/text/utils/static-extract-text'
 import {
   AnyEditorDocument,
   EditorExerciseDocument,
   EditorRowsDocument,
   EditorTemplateExerciseGroupDocument,
 } from '@editor/types/editor-plugins'
-import { isRowsDocument } from '@editor/types/plugin-type-guards'
+import {
+  isMultimediaDocument,
+  isRowsDocument,
+} from '@editor/types/plugin-type-guards'
 
 import { InstanceData } from '@/data-types'
 import { getTranslatedType } from '@/helper/get-translated-type'
@@ -27,4 +30,25 @@ export function getPreviewStringFromExercise(
   return `${typeString}: "${
     titleString.length < 60 ? titleString : titleString.substring(0, 50) + '…'
   }"`
+}
+
+export function extractStringFromRowsTextAndMultimedia(
+  document?: AnyEditorDocument
+): string {
+  if (document) {
+    if (isRowsDocument(document)) {
+      return document.state
+        .map(extractStringFromRowsTextAndMultimedia)
+        .join(' ')
+        .trim()
+    }
+
+    if (isMultimediaDocument(document)) {
+      return extractStringFromRowsTextAndMultimedia(document.state.explanation)
+    }
+
+    return extractStringFromTextDocument(document)
+  }
+
+  return ''
 }
